refactor(array): モジュールのインポートパスを整理し、ファイル構造を改善
