+++ package/main/src/Array/chunk.ts
@@ -1,4 +1,4 @@
-import type { ChunkArrayType } from "$/arrayType";
+import type { ChunkArrayType } from "$/array/chunk";
 /**
  * 配列を指定した数で分割する
  * @param arr 配列
