+++ package/main/src/tests/types/arrayType.test.ts
@@ -1,4 +1,4 @@
-import type { ChunkArrayType } from "@/types/arrayType";
+import type { ChunkArrayType } from "$/array/chunk";
 
 describe("arrayType", () => {
   it("ChunkArrayType", () => {
