+++ package/main/src/types/array/chunk.ts
@@ -1,4 +1,4 @@
-import type { Length } from "./logicType";
+import type { Length } from "$/logicType";
 
 type Chunk<
   T,
@@ -16,13 +16,3 @@ type Chunk<
     : [...R, C];
 
 export type ChunkArrayType<T, N extends number> = Chunk<T, N>;
-
-type ZIP<T, O extends unknown[] = []> = T extends [infer Head, ...infer Tail]
-  ? Head extends unknown[]
-    ? ZIP<Tail, [...O, Head[number]]>
-    : never
-  : T extends []
-    ? O[]
-    : T;
-
-export type ZipArrayType<T extends unknown[][]> = ZIP<T>;
