+++ package/main/src/types/array/index.ts
@@ -1,0 +1,2 @@
+export * from "./chunk";
+export * from "./zip";
